+++ src/services/interfaces/IBaseService.ts
@@ -1,4 +1,4 @@
-export interface IService<T> {
+export interface IBaseService<T> {
   findAll(): Promise<T[]>;
   findOne(id: number): Promise<T | undefined>;
   create(entity: T): Promise<T>;
